moved config
